type

+++ 0x00-hello_world/6-size.c
@@ -1,15 +1,15 @@
 #include <stdio.h>
 /**
- * main - a program that print the size of various types on computer
+ * main - a program that print size of various types on computer
  *
  * Return: Always 0 (Sucess)
  */
 int main(void)
 {
-	printf("Size of char: %lu byte(s)\n", sizeof(unsigned char));
-	printf("Size of int: %lu byte(s)\n", sizeof(unsigned int));
-	printf("Size of long int: %lu byte(s)\n", sizeof(unsigned long int));
-	printf("Size of long long int: %lu byte(s)\n", sizeof(unsigned double long int));
-	printf("Size of float:%lu byte(s)\n", sizeof(unsigned float));
+	printf("Size of char: %i byte(s)\n", sizeof(unsigned char));
+	printf("Size of int: %i byte(s)\n", sizeof(unsigned int));
+	printf("Size of long int: %i byte(s)\n", sizeof(unsigned long int));
+	printf("Size of long long int: %i byte(s)\n", sizeof(unsigned long long int));
+	printf("Size of float:%i byte(s)\n", sizeof(unsigned int));
 	return (0);
 }
